finagle-core: Better OffloadFilter position (under a toggle)
Problem

Today's OffloadFilter is sitting on the edge between the application and the Finagle stack, forking execution
just as it enters the userland. This approach implies a suboptimal near-CPU performance as most of the data
needed to handle a request won't be cache-ready.

Solution

Fork execution earlier on the edge between the Netty pipeline and the Finagle stack. This way all data required
by the application logic will be waiting in the cache. The new position is placed under a toggle (disabled by default)
for a smooth and gradual roll out.

Result

Application with offload filter enabled just became a little faster. Forking into the application thread earlier also
means there is less work to be done by the IO thread, which presumably means the Netty worker pool can be shrunk
even further (if needed).

JIRA Issues: CSL-10319

Differential Revision: https://phabricator.twitter.biz/D604410