util-routing: evolve routing result beyond Option[Route] and simplify Router generation
Problem

A `c.t.util.routing.Router` represents the result of finding
a `Route` as a `Option[Route]`. As we implement new
use cases around the `Router`, we have found a need
to pass along the routing `Input` as part of the result.

Solution

Introduce a new `c.t.util.routing.Result` type for representing
a `Router` result, which can be represented by different
states: `Found` and `NotFound`. As a side-effect, we have
the ability to return a result that represents a router
being closed instead of being forced to throw an exception
for states that do not match Some or None. This also
simplifies our Java-friendly story and allows us to remove
`AbstractRouter`.

It was also found that the `Generator` should use an
encapsulating type for its input instead of a Tuple, which
has also been reflected in these changes.

Result

The result of routing via a `Router` can pass on more
information than was previously possible. Generating a
`Router` is simplified.

GitOrigin-RevId: 2741ba90522c6f21f882e4c710256532e24822d4