finagle-core: introduce request and response wrappers for experimental RoutingService
Problem

When dealing with routing a reqeust, there may be extra information either
extracted as part of the routing process (i.e. HTTP Path parameters) or required
to route (i.e. Thrift Method). We were previously encapsulating this behavior as
part of the `RequestTransformingFilter`, but this abstraction is not clear and
also does not allow for access to the request as part of route determination.

Solution

Let's introduce a `Request` and `Response` container for the respecitve `Req`
and `Rep` types of the `RoutingService`. By making a `Route` act over the
`Request` and `Response` types, we can inject/annotate fields for different
concerns at different phases of execution (i.e. filters before/after matching,
the router during the matching process). The `Field` annotations are a simple
API inspired by `http.Request.Schema.Record` and also avoids the need for API
extenders to introduce or know about special Locals. The Request/Response and
Field API's should also be considered experimental/unstable, but allows us to
continue to iterate on our HTTP routing implementation.

JIRA Issues: CSL-10468

Differential Revision: https://phabricator.twitter.biz/D610614