finagle-mux: Add API for Request Which Only Needs the Body
Problem / Solution

When testing functionality via the Mux protocol, if only testing the
internal mechanics of how clients and servers interact, specifying a
`Path` for a mux `Request` to use is unnecessary and adds to extra code
to write and maintain. This change provides the ability to create a mux
`Request` in both Scala and Java without specifying a `Path` explicitly.
In those cases `Path.empty` will be used.

Differential Revision: https://phabricator.twitter.biz/D613686