finagle-core: allow applications to reuse the offload filter thread pool
**Problem**

Some applications need to run tasks in parallel but currently it's necessary to create a new thread pool to achieve that. Creating a separate thread pool is problematic in containerized environments since it can lead to CPU Throttling and is difficult to tune.

**Solution**

Expose a method that allows the application to reuse the offload thread pool to execute tasks.

Differential Revision: https://phabricator.twitter.biz/D606438