finagle-mysql: Added new embedded MySql utilities for integration testing.
Problem

finagle-mysql currently runs its integration tests on pre-configured
environments. This requires a running and configured mysqld instance on the test
runner which is not ideal.

Solution

Introduce a testing harness which bootstraps the mysqld instance programatically
when given the path to the respective mysqld binary. The package is integrated
into ScalaTest via fixtures and allows an integration tests/suite to easily make use
of the boostrapped instance. For example:

```
  test("ping default") { fixture =>
    val roClient = fixture.mySqlDatabase
      .createROClient().newRichClient(fixture.mySqlInstance.dest)
    Await.result(roClient.ping(), Duration.fromSeconds(1))
  }
```

Result

Running our integration suite in various environments only
require pointing to the respective mysql extracted binaries
location.

Differential Revision: https://phabricator.twitter.biz/D620448