finagle-http: update experimental HttpRouter to route via Request.path
Problem/Solution

The experimental HttpRouter is using the `c.t.finagle.http.Request.uri`
for doing route determination, which includes the query params. We
really want to be using the `c.t.finagle.http.Request.path` method, which
does not include the query params as part of path matching. It's also
worth noting that the `c.t.finagle.http.Request.path` method will calculate
the path each time it is called, so we want to memoize that as the path is
checked often. Let's update our logic and add tests.

JIRA Issues: CSL-10736

Differential Revision: https://phabricator.twitter.biz/D624699