twitter-server: add histogram bucket metric names to metric_metadata.json
Problem
Consumers of the metric_metadata.json endpoint do not have a good way of discovering the metrics which make up a histogram entry in metric_metadata.json.

Solution
Replace the array type bucket field of metric_metadata.json (which simply listed the percentiles) with an Object field mapping from names (min, max, etc) and percentiles to the suffix for the metric name representing that portion of the histogram.

Result
The metric_metadata.json output now has a map from percentiles et al to the suffix for the metric name for that percentile (or max, min, etc).

JIRA Issues: CSL-10261

Differential Revision: https://phabricator.twitter.biz/D627870