finagle-mux: ensure PushSessionQueue gauge gets cleaned up
Problem/Solution

The `PushSessionQueue` "negotiating_queue_size" Gauge does
not get cleaned up until a GC occurs, due to the WeakReference
behavior on `CumulativeGauge`s. This adds overhead in calculating
the Gauge's value between GC's, which may not occur frequently.
This behavior seems to be exacerbated when the Aperture
EagerConnection feature is in use with the ForcedWithDtab setting.

Let's clean up the Gauge when it is no longer needed.

JIRA Issues: CSL-10774

Differential Revision: https://phabricator.twitter.biz/D632571