finagle-core: use EndpointFactory in place of AddressedFactory where the address is not reassigned
Problem:
`AddressedFactory` takes two parameters: an `EndpointFactory` and an `Address`.
However, address should always be the same as factory.address. TrafficDistributor
is the only place this functionality is used (by replacing an `Address` with a
`WeightedAddress`). In all other places, an `EndpointFactory` should be sufficient

Solution:
Add lazy `weight `value to `EndpointFactory` to reduce the overhead of unpacking
weight from Addresses and replace AddressedFactory where possible

JIRA Issues: CSL-10602

Differential Revision: https://phabricator.twitter.biz/D614117