upgrade netty to incorporate new aarch_64 classifier
Problem
This is a refresh of #884: "There are new netty releases and the
classifier for AArch64 was changed from linux-aarch64 to
linux-aarch_64. This results in dependency problems with other
projects."

Solution
Upgrade to netty 4.1.59.Final and netty-tcnative 2.0.35.Final
for security fixes and AArch64 dependency rename.

#884 was blocked because of a bug in netty 4.1.56.Final, but
AFAICT this has been fixed in 4.1.59.Final.

Signed-off-by: Jing Yan <jyan@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D629268