finagle/doc: Update link to Zookeeper ServerSet
Problem/Solution

twitter/commons was archived and now resides under
twitter-archive/commons. As such we also no longer
generate javadocs for the library. However there
is still a link in the Finagle user guide, Names
section which references the non-existent javadoc.

Update the link to point to the source file
directly. This isn't perfect but it is better than
a broken link.

Differential Revision: https://phabricator.twitter.biz/D636633