util-stats: label expressions with service name
Problem

Expressions with the same name clobber each other inside of the Metrics store.

Solution

Use RoleConfiguredStatsReceiver to store the name of the Client/Server and use this value to populate the ServiceName label, which is used in turn as part of the expressions name in the internal Metrics storage.

Result

Multiple success_rate queries for different portions of a service can now coexist.

Differential Revision: https://phabricator.twitter.biz/D630620