{util-core,finagle-core}: Make it possible to directly construct X509Certificates without an on-disk file
Problem

Right now our entire X509Certificate chain is File-based.  This doesn't work
well in cases where you don't want to have to use a File.

Solution

Add a new way to construct an X509Certificate and TrustManager that are based
on a buffered String, instead of File.  Also rename PemFile to PemByte so that
it assumes being fully-buffered instead of not-yet read.

JIRA Issues: CSL-10763

Differential Revision: https://phabricator.twitter.biz/D641088