finagle-core: change fallback behavior in deterministic aperture to be effectively p2c
Problem:
We're redesigning deterministic aperture so that it understands weights.
Random aperture will remain unaware of weights, so it doesnt make sense
for random aperture to be our fallback.

Solution:
If process coordinates are unset, use deterministic aperture but pretend
the server has no peers. This will include all endpoints, effectively
resorting to p2c aperture but retaining the ability to move back to
deterministic aperture if process coordinates are reset.

While we're at it, let's modify `dapertureActive` so that this fallback
behavior has meaning (as is, it's near impossible to reach the fallback case)

JIRA Issues: CSL-10604, CSL-10573, CSL-10642

Differential Revision: https://phabricator.twitter.biz/D637317