finagle-core: Introduct 'IgnoreIRTEs' ResponseClassifier
Problem

There are legitimate scenarios where a server will have a Finagle client with
a super low `RequestTimeout`. Sometimes, this is because the retrieval of a
response is seen as fortuitous if it arrives in time, but not a problem if it
doesn't. In this scenario, an `IndividualRequestTimeoutException` is not an
indication of `Failure`, but only an indication of non-success. Treating an
`IndividualRequestTimeoutException` as a `Failure` in this scenario can lead to
problems with a circuit-breaker like 'Failure Accrual', as servers are harshly
penalized when there's not an actual problem with them.

Solution

Add a new `ResponseClassifier`, 'IgnoreIRTEs', which handles all responses the
same as the 'Default', except for `IndividualRequestTimeoutException`s. For
'IRTEs', this `ResponseClassifier` treats them as `ResponseClass.Ignored`.

JIRA Issues: SSTR-5245

Differential Revision: https://phabricator.twitter.biz/D645818