finagle-core: service loaded inet resolver support
# Problem

Finagle's default DNS resolution uses the JDK blocking resolvers. The use of a thread pool to work around the blocking call is problematic since creating more threads can generate CPU throttling in containerized environments and, even if the concurrency is limited (which is the case for the inet resolver), the queue of pending requests is unbounded and can create more GC pressure.

# Solution

This phab introduces the ability to use a service loaded inet resolver, which is currently being implemented in a separate module.

JIRA Issues: VM-4490

Differential Revision: https://phabricator.twitter.biz/D644533