metrics: Add kernel for memoizing the metric builders
Problem

Metric metadata including fully scoped names is filled through underlying
StatsReceiverProxies. This means the same MetricSchema we passed in Expressions
doesn't have any metadata.

Solution

For metrics we used to build expressions, we hash the MetricBuilder and memorizing
in the Metric storage using the object reference hashCode as hashMap key. For
registered expressions, replace all MetricBuilders with the fully hydrated ones.

Result

This is a workaround solution to buy us sometime re-design the StatsReceiver to carry
MetricBuilder information so we don't need to back-filling.

JIRA Issues: CSL-10845

Differential Revision: https://phabricator.twitter.biz/D645590