finagle-core: Use pending probe tasks for admission control
Problem

When using the signal as to whether the queue is empty or not
we have to average for a while because the queue having work is not
by itself bad, it's if it never empties. This averaging makes the response
time longer.

Solution

Instead of using queue depths use queue latency, with a fun twist. Instead
of putting a task in and seeing how long it takes we continually put tasks into
the queue at a fixed rate (1ms) and if there are N tasks outstanding it means
there is roughly a N millisecond latency to the queue. It turns out this is
extremely responsive since the moment the queued tasks are completed
admission control stops. It's also an early predictor instead of a late predictor
because if you have to wait until the task has finished to know the latency
you may not know until after the latency event is over.

This method appears to be far simpler and more responsive both in terms of how
fast it can detect latency events (based on much smaller window) and in terms
of rejecting work only when we're _currently_ in a latency window and immediately
recovering once we exit the latency period instead of needing to wait for probe
tasks.

Differential Revision: https://phabricator.twitter.biz/D647875