util-stats: Add metadata support to metrics instances
Problem

We want to construct metrics expressions, but it's challenging without having
access to the fully constructed `MetricBuilder` instance.  We don't have access
to the fully constructed instance because of the way StatsReceiver composition
works.  We only have partial context when we actually invoke the MetricBuilder,
and the rest of the context is assembled when the metric is constructed.

Solution

We should provide access to the fully constructed `MetricBuilder` instance after
the metric has been constructed.  We also introduce a `Metadata` abstraction to
represent when the metric actually represents multiple metrics behind the scenes.

JIRA Issues: CSL-10887

Differential Revision: https://phabricator.twitter.biz/D653751