finagle-core: add ability to configure custom field information to Route
Problem

As we look to integrate our experimental HttpRouter in different
areas of the Twitter-stack, we need to be able to associate
custom route information that can be used for customizing service
behavior. An example of this is some of the extra UI index information
for the `c.t.server.AdminHttpServer.Route`. Another example is in
Finatra, which handles Admin concerns, but also has extra information
for display in the registry.

Solution

Let's make our `Field` more generally usable for not just a `Message`,
but also for our `Route`. We refactor our map lookup behavior so that
we can have a shared `FieldMap` for use by our routing structures.

Result

We will better be able to support integrations in the Twitter-stack.

JIRA Issues: CSL-10113

Differential Revision: https://phabricator.twitter.biz/D625077