finagle-mux: Better test for whether RemoteInfo was added
Problem

One of the tmux tests tries to test what happens when a
BadMessageException comes through by making a bad server implementation.
This is unnecessary and may be the source of some flakiness.

Solution

Delete the bad server implementation and just make a bad ServiceFactory
and stick it in as the endpointer. This avoids a socket connection altogether and
still tests the default client stack.

JIRA Issues: CSL-10921

Differential Revision: https://phabricator.twitter.biz/D661701