finagle-netty4: Fix flaky Netty4ListenerTest
Problem

When testing whether we can limit the number of connections we set the
max to 1 connection for the server and then try to open two different
connections. We attempt to enforce order by writing a byte to client 1.
This doesn't appear to be working. This is likely because although we
establish a connection and synchronously write + flush a byte, that
doesn't necessarily mean we've even established a connection yet, just
that we've put the bytes in a buffer and asked the OS to send packets.

Solution

Instead let the server tell us when the connection has been established
through its metrics, via the `"connections"` gauge. Only after the server
acknowledges the first connection should we try to establish the second.

JIRA Issues: CSL-10921

Differential Revision: https://phabricator.twitter.biz/D662406