finagle-mysql: Add params and configs for caching_sha2_password
Problem

As part of the next step of implementing the `caching_sha2_password`
authentication method, we need a way to enable it. We also need a way to
configure the location of the server's RSA public key in addition to enabling a
cache miss to force full authentication when testing to ensure we are
authenticating correctly.

Solution

Add params and methods to configure the client to 1) enable
`caching_sha2_password` auth, 2) configure the local path to a copy of the
server's RSA public key, and 3) enable a cache miss to force full
authentication.

JIRA Issues: CSL-10683

Differential Revision: https://phabricator.twitter.biz/D665745