finagle-http: make handling of invalid URI consistent across client implementations
Problem

We have inconsistent behavior when faced with a an illegal URI
across our HTTP/1.x and HTTP/2 client implementations. For the
HTTP/2 client, this results in unhandled exceptions propogating
as `UnknownChannelExceptions` when the `java.net.URI` parser is
called under the covers on invalid input via the Netty pipeline.
For HTTP/1.x clients, we see the clients happily able to submit
a request with invalid input. We should really see a more consistent,
non-retryable exception get thrown in all cases, while still ensuring
that our server implementations consistently respond to these requests
with a `400 Bad Request` status.

Solution

Add better error handling to the `UriUtils`, which is used across
our implementations to handle invalid escape sequences. Additionally,
the error handling does not behave as expected for pure HTTP/2
clients, which do not have any validation/mapping installed in their
pipelines. Let's add tests and ensure that we install this handling
more consistently and correct this behavior through better test
coverage. We will now throw a `InvalidUriException` when attempting
to submit a request with an invalid URI and ensure this exception
is marked as non-retryable.

Result

More consistent behavior across HTTP client implementations.

JIRA Issues: CSL-10770

Differential Revision: https://phabricator.twitter.biz/D660069