finagle: Support TLS Snooping in Mux
Problem

We use thriftmux all over the place and it supports downgrading to vanilla
thrift so that non-thriftmux clients can still play in our ecosystem. However,
after we added opportunistic tls and turned it on everywhere that is no
longer true since we need to make sure the connection is secure and we
previously only allowed that through a cleartext-upgrade, only supported
in mux.

Solution

Support TLS snooping in finagle mux. If a client starts a connection as TLS
we consider that the same as the client saying TLS is required. This allows
us to downgrade to vanilla thrift via the same path we'd normally follow with
cleartext and still allowing the server to require a secure connection.

Differential Revision: https://phabricator.twitter.biz/D584638