finagle-mysql: Migrate more integration tests to the testing harness
Problem

We have two ways to run our integration test, one is with the IntegrationClient,
and the other is the embedded mysql harness. The harness is a better solution,
so we should migrate the existing tests to use the harness.

Solution

Migrate the AbortedClientTest and the TransactionTest suites. Also, remove
some override statements as they are unnecessary.

JIRA Issues: CSL-10990

Differential Revision: https://phabricator.twitter.biz/D671974