finagle-mysql: Add EmbeddedSimpleSuite to the testing harness
Problem

We have two ways to run our integration tests, one is with the
IntegrationClient, and the other is the embedded mysql harness. The harness is a
great solution, but the way the `EmbeddedSuite` is set up, it is hard to write
tests that use a single result from the db without making the query to the db
during every test.

Solution

Add the `EmbeddedSimpleSuite` that initializes and exposes the `Fixture` during
test suite initialization and migrate two test suites.

JIRA Issues: CSL-10991

Differential Revision: https://phabricator.twitter.biz/D672184