finagle-zipkin-core: Don't unnecessarily hold onto Locals in DeadlineSpanMap
Problem

We use a timer in DeadlineSpanMap to wait for a period of time before
logging spans. Unfortunately the timer unnecessarily captures the Locals
as well, and these can be pretty heavy in certain cases, resulting in
bad GC behavior.

Solution

Clear out the locals before setting the timer. There is nothing in the
locals that we need at that point so we should be able to drop them.

JIRA Issues: CSL-11038

Differential Revision: https://phabricator.twitter.biz/D685929