finagle-memcached: Remove Ketama Partitioned Client
Problem

The Partition Aware Client has introduced to Memcached a while ago and coexisted with the KetamaPartitionedClient, and the latter creates a Twemcache client for each endpoint, which adds extra pressure to the cluster discovery service.

Solution

Most of the customers/clients have migrated to using the cluster discovery service and we can finally get rid of the old Memcached client.

Result

  # Although might be negligible, moving the current old-style Memcached clients to partition aware client can remove some request pressure to the cluster discovery service,
  # Remove redundant code and have a clean codebase to add new features.

JIRA Issues: CSL-10701

Differential Revision: https://phabricator.twitter.biz/D661460