finagle-mysql: Migrate client and ssl tests to embedded harness
Problem

We have two ways to run our integration test, one is with the IntegrationClient,
and the other is the embedded mysql harness. The harness is a better solution,
so we should migrate the existing tests to use the harness.

Solution

Migrate the ClientTest and MysqlSslTest suites.

JIRA Issues: CSL-11053, CSL-11054

Differential Revision: https://phabricator.twitter.biz/D690943