finagle-mysql: Remove the IntegrationClient and simplify EmbeddedSimpleSuite
Problem

Now that all of the mysql integration tests are migrated to the embedded testing
harness, the IntegrationClient is unused. Additionally, @ibennett correctly
pointed out that there is some repetition of code which should be fixed.

Solution

Remove the IntegrationClient, and add a run method to the EmbeddedSimpleSuite
that takes a lamda and runs it.

JIRA Issues: CSL-11074

Differential Revision: https://phabricator.twitter.biz/D693049