finagle-http2: Do some pre-checks when opening a HTTP2 stream
Problem

The opening of streams can race with go-away and other stream openings
and find itself in a state where we violate the max concurrent streams
or opening a stream during a drain. This can lead to ugly errors for
users.

Solution

Do some pre-checks as part of the initial stream opening. We enter the
netty executor responsible for the channel and thereby avoid any race
conditions and emit Failures which should be able to be retried.

JIRA Issues: CSL-9430

Differential Revision: https://phabricator.twitter.biz/D712424