finagle: Fix metric documentation for BackupRequestFilter
Problem

In code and in docs the "backups_won" counter for BackupRequestFilter is documented
as being the number of requests that completed before the original, regardless of if they
were successful. This is incorrect; the counter is actually for the number of requests
that completed *successfully* before the original.

Solution

Correct the documentation in code and in the general metrics documentation.

Differential Revision: https://phabricator.twitter.biz/D716076