finagle-core: Order the nodes in the aperture additional metadata
Problem / Solution

We currently use a Set to add the collection of nodes to the additional
metadata and that doesn't preserve order very well. When comparing
instances it's nice to be able to go down the list in order.

Solution

Order them by turning it into a sequence and sorting.

Differential Revision: https://phabricator.twitter.biz/D722030