finagle-netty4-http: allow setting multiple values for a request cookie
In a Response, CookieMap allows setting multiple cookie values for a given name as long as they differ by their domain or path. However, in a Request, if the client sent several values for a given cookie name, only the last one is currently available.

This fixes a bug when decoding request cookies, and changes CookieMap behavior to allow inserting multiple values for a request cookie. It also preserves insertion order for the different values for a given cookie name, to preserve the order in which the client sent the values in a request.