finagle-core/partitioning: Close balancers when repartitioning
Problem

We discovered that when partitioning is enabled, gauges were being leaked. We
weren't closing the gauge-containing balancers before because 1) it would close
the endpoints, and 2) we thought any un-closed resources were low-cost (this was
not true).

Solution/Result

Add a stack param that informs TrafficDistributor (where balancers and endpoints
are closed) to close only balancers when partitioning is enabled. This, in turn,
closes the costly gauges.

JIRA Issues: CSL-11271

Differential Revision: https://phabricator.twitter.biz/D731675