finagle-postgresql: Propagate request timeouts to pending readers
Problem:
If a request timeout is set, the timeout only applies until the service returns a
response.  However, for streaming queries, the response is not complete until the
returned reader(s) are complete.

Solution:
Rather than using the built-in TimeoutFilter, use add logic at the client level to
propegate timeouts to the returned readers.  Additionally, handle the interrupt
in ExecuteMachine and use it to close the current connection, as it is no longer
valid.

JIRA Issues: CSL-11315

Differential Revision: https://phabricator.twitter.biz/D743065