[finagle-postgres] Configurable cancel grace period on interrupt.
Problem:
When a query is interrupted a ClientDispatcher closes the connection in order to cancel the query. This results in a connections churn if number of cancellations is
significant. This can be a problem even in a healthy state, where there is a constant stream of cancellations due to Backup Requests and/or other causes of
client request abandondment.
Excessive connection churn can result in a spiral of death, where connections can be closed faster then new connections sucessfully created.

Solution:
Instead of immediately closing connection after cancellation allow some grace period for the query to complete on its own, and close the connection
only when grace period expires. This will reduce the number of connection closes for queries that complete within reasonable time and/or have a server-side statement timeout.

This change adds a CancelGracePeriod parameter to the finagle-postgres client, with default timeout of 0ms. If configured, the connection will only be closed after
the grace period expires once an interrupt is received.

The new cancellation logic and grace period applies to the entire dispatch action, which includes both returning of the initial Response, as well as processing of that Response (for example processing of the returned Reader(s)).
Entire dispatch action will be given opportunity to complete gracefully and restore connection to Ready state, as long as the grace period doesn't expire.
This process works also in case when a reader was already closed, in which case the intermediate data backend messages will be discarded.

This approach is also compatible with potential future extension, if a cancel request were to be sent to a server.

Differential Revision: https://phabricator.twitter.biz/D747930