[finagle-postgres] Include NackAdmissionFilter in the client stack.
Problem:
In scenario where destination database server/cluster has degraded success rate and/or latency finagle-postgres will continue sending traffic. This results in even higher number of outstanding queries and higher parallelism that further degrades the database/cluster and prevents recover even after initial trigger is removed.
Client-side NackAdmissionFilter is a good candidate of a module that would address that problem, however NackAdmissionFilter module is finagle-private and is not included currently in the finagle-postgres client stack. This makes it difficult/awkward for users of finagle-postgres to configure NackAdmissionFilter.

Solution:
- Include NackAdmissionFilter in the finagle-postgres stack, so that it can be configured for large scale distributed case.
- Keep NackAdmissionFilter disabled by default, for backwards compatibility, and as a reasonable default when using finagle-postgres in smaller scale settings.

Alternative solution could be to make NackAdmissionFilter.module public to allow for easier configuration by clients what want to add this filter to the stack. The  drawback of this approach is that it would expect each clients to properly choose the position in the stack where the module should be added.

Differential Revision: https://phabricator.twitter.biz/D760520