[util-core][fiber-scheduler] remove async semaphore mechanism
**Problem**

During the development of the Fiber Scheduler, an async semaphore was used as an attempt to keep the concurrency behavior of tasks coming from redirected `FuturePool`s. The idea is that the async semaphore would limit the concurrency of tasks similarly to how the number of threads in a `FuturePool` limits the number of concurrent tasks.

This approach is incorrect because a `FuturePool` limits synchronous concurrency and the async semaphore mechanism limits asynchronous concurrency, which has completely different behavior.

**Solution**

The async semaphore got introduced with potential future issues in mind and hasn't been proved necessary yet. Given that and the fact that a creating a mechanism to limit synchronous concurrency would require a more delicate change, this phab removes the concurrency limitation mechanism.

Differential Revision: https://phabricator.twitter.biz/D762557