finagle-core: Fix failure to interrupt infinite resolutions
Problem

If you attempt to resolve an address that just doesn't exist using an
infinite backoff stream you end up with a never resolving Future. That's
fine, but the Var that is intended to be updated by that future will
never interrupt the resolution and thus that interest will live forever.
That would also be fine, but it can't get GC'd because that interest may
end up living in a cache or even just by being perpetually bounced through
timer threads.

Solution

Use the Closable correctly with Var.async and have that Closable
raise on a pending future if the future is still outstanding. This will
properly prune interests from wait queues.

Differential Revision: https://phabricator.twitter.biz/D759466