The test target "//finagle/finagle-core/src/test/scala:scala" is failing too often in master in CI.
Summary:
Problem:

The test target "//finagle/finagle-core/src/test/scala:scala" is failing too often in master in CI.

Logs available here:

Known to fail ticket: [CSL-11462](https://jira.twitter.biz/browse/CSL-11462)

Solution:

It has been marked "known-to-fail" in the BUILD file

Result:

This will cause it to be skipped in future builds, and its failures ignored in submit queue. This implies that code changes that further break this test will not be prevented from being submitted to the master branch, until the problem is fixed and the test is no longer marked known-to-fail.

TBR=true

Reviewers: O847 source:/finagle/finagle-core/!

Subscribers: O46742 source:/finagle/finagle-core/

Differential Revision: https://phabricator.twitter.biz/D781138