finagle-memcache: Add public methods for gracefully testing validity of memcache keys
Problem

The only way users can check the validity of a memcache key is to use it and
have the API throw an exception if it is invalid. A graceful method for early
validation of the keys is desired.

Solution

Provide public methods BaseClient#validKey and JavaClient#isValidKey that
execute the same validation logic as the rest of implementation, but gracefully
return a boolean value instead of throwing an exception.

JIRA Issues: CSL-11448

Differential Revision: https://phabricator.twitter.biz/D782660