finagle|finatra|scrooge|ts|util: Add build version to GH Actions build ccache key
Problem

Some generated files are cached in the build cache but the thrift files have
since been removed, which causes build failures.

Solution

Add a way to easily invalidate the build cache so if we experience something
like this again, we can simply bump the version number and move on to a
successful build.

Differential Revision: https://phabricator.twitter.biz/D792859