finagle-zipkin-core: Increase sample rate precision
Summary

Problem

When a Zipkin Sampler is configured to sample at a rate lower than 1/10,000 (but
not zero), it will sample at rate of 1/10,000. This is an issue in a
high-traffic setting where a lower value would be needed.

Solution

Internally the decision for whether to trace a request or not was made by
multiplying the rate by 10,000 and comparing it to a random int value derived
from request id modulo 10,000. If a lower rate was configured the muliplied rate
would be a float value between 0-1, which would still select int values n for
which n % 10000 == 0. The choice of 10,000 was arbitrary; a larger multiplier
could be safely chosen. 32-bit floats can represent all integers up to
16,777,216 without loss of precision so it was chosen as the new multiplier to
be used for highest fidelity.

Result

Sampler will now produce correct sampling for configured rates as low as
1/16,777,216.

Differential Revision: https://phabricator.twitter.biz/D792693