finagle-mux: Lazily initialize shared mux stats
Problem

We are eagerly instantiating the shared mux stats in the mux client and
server. This is not good because the stats are not appropriately scoped
until after the client or server are initialized.

Solution

Make them lazy so that they're not instantiated until we actually need
them, at which point the stats receiver should be appropriately scoped.`

Result

This change drops 5 counters (that are always zero) from each JVM instance.

JIRA Issues: CSL-11527

Differential Revision: https://phabricator.twitter.biz/D801250