[finagle-mysql] EmbeddedDatabase cache refinements
Problem:
There exists a usability issue the the initialization of an embedded
database. Currently databases are cached by name and destination only. This
creates the scenario where different test suites could be asking for a database
with a common name but different users or setup queries. The code today will
return the previously configured database that matches the name. This results in
a non-deterministic state of the retrieved database.

Solution:
The database cache has been enhanced to check the users created on an
instance, along with the setup queries used to setup the databases. The cache
works as follows.
If the database name has not been created, configure normally.
If the database name has been created, ensure the Users passed in with the
configuration have either never been created, or were created with the
same passwords. The latter check is to ensure passwords do no change during the
execution of test suites.
If the database name has been created and either the Users have different
passwords already configured OR the setup queries are different than the
initial configuration, throw an exception.

Differential Revision: https://phabricator.twitter.biz/D804085