[finagle-mysql]: Loosen restrictions on --socketFile server parameter
Problem:
In sandbox environments, there can be scenarios where the tmp
directory is very deep which could result in socket files
with names greater than the maximum allowed by operating
systems. This causes the mysql server to fail to start
and results in failed tests.

Solution:
Remove the restriction where --socketfile isn't allowed
in the configured server start parameters defined for a
Mysql version.  If --socketfile isn't provided, the library
will behave as it did before this change by creating a
temp file in the extracted Mysql directory.

Differential Revision: https://phabricator.twitter.biz/D805375