finagle-core: close resources when avoiding TrafficDistributor
Problem:
The weighted load balancers were built to avoid using
the TrafficDistributor. Unfortunately, the TrafficDistributor
had been responsible for closing resources in other balancers.
This resulted in connections not closing properly when using
Weighted balancers

Solution:
In the close() method for all Aperture balancers, we'll add in
conditional logic that closes resources if we're using weighted
balancers

JIRA Issues: CSL-11577

Differential Revision: https://phabricator.twitter.biz/D808684