finagle-core: Set panic mode for load balancers
Problem:
Panic mode is when the load balancer fails to find a node with `Status.Open`
after multiple attempts and fails open. The request is sent to the last picked
node, even if it is unhealthy. Panic mode exists in Finagle, but it is
obscurely configured by `maxEffort`, a fixed number of LB retries. It is hard
to know what percent of replica set can be unhealthy before panic mode
starts.

Solution:
Introduce `LoadBalancerFactory.PanicMode` stack param. This configures the
percentage of nodes in the replica set (size of the aperture) that can be unhealthy
before panic mode kicks in. This percent  is an approximation because of rounding
and the randomness from probabilistic load balancers.

Result:
At most 1% of the requests will enter panic mode when the percent of unhealthy
nodes is at the threshold. This provides a map from "percent unhealthy
threshold" to "max effort". The default is kept at maxEffort=5, which is
PanicMode.VeryUnhealthy.

JIRA Issues: CSL-10896

Differential Revision: https://phabricator.twitter.biz/D779468