finagle-core: maybe remove an intermediate collection when constructing a Drv
Problem

It turns out that the c.t.f.util.Drv class makes quite a bit of garbage.

Solution

Eliminate some of that garbage by removing an intermediate collection.
If we construct a Drv via un-normalized weights we normalize via an
intermediate collection. Instead, we can just normalize in place by
using an Array instead of a Seq.

Differential Revision: https://phabricator.twitter.biz/D819073