scrooge: generate a method to validate request and return the violations in Scala
Problem/Solution
For each method defined in a `service` in the Thrift IDL, generate a new
API `violationReturning<method_name>` in `MethodPerEndpoint` to
validate incoming request (of Struct, Union, Exception types) by invoking
the `ValidateInstanceValue` method on the request, and return any
violations (as method parameters `<request_variable>Violations`)
back to the users in the method API, the new method will by default
wrapping the implementation of the existing method.

Scrooge will provide the implementation for the new method for client,
by wrapping the implementation of existing method, there is no behavior
change on the client side.

When there is validation annotation on the IDL, finagle will call the new
`violationReturning<method_name>`, if user doesn't implement the new
method, it will be directing to the existing method; if user implement the
new method, the existing behavior will be overriden by the new
impementation.

**Added a flag `hasValidationAnnotation` in the parser, all new APIs
will only be generated when the flag is true. There is no runtime
behavior change on existing services.**

Result
When `violationReturning<method_name>` is unimplemented, both
call to `<method_name>` and `violationReturning<method_name>`
will use the implementation of `<method_name>`, and throw an
exception when any request fails validation.
When `violationReturning<method_name>` is implemented, both
call to `<method_name>` and `violationReturning<method_name>`
will use the implementation of `violationReturning<method_name>`,
where no exception is thrown per request validation violations.

JIRA Issues: CSL-11150

Differential Revision: https://phabricator.twitter.biz/D799294